Make 'python setup.py test' work
Packagers sometimes like to use 'python setup.py test'. This
change adjusts setup.py and setup.cfg to allow this to work
with pytest instead of nose. To make things work, the
settings for coverage need to be moved into tox.ini.

A SkipTest in test_config_middleware is changed to its pytest
equivalent.